requirements.txt
tensorflow_gpu
tensorflow_addons
scikit-learn
matplotlib
seaborn